Test session cleanup when expired (2/n)
Note mStagingDirectoryRule is slightly tweaked.

The staging directories left by the previous test cases will be
deleted when we fast-forward system clock and run expiration. We
can no longer assert that stagedAfter is identical to stagedBefore.
Instead, we have to assert that stagedAfter is a subset of
stagedBefore to make sure staging directories created during tests
are correctly deleted.

Bug: 209551994
Test: atest CtsRootPackageInstallerHostTestCases
Change-Id: I35e37136184a0006135a3a78312e03813741e6ba