Fix CtsRootPermissionTestCases NoSuchMethodError before U.
We fixed the test issue on U by skipping it when SdkLevel.isAtLeastV()
returns false. However, despite that the SdkLevel is a class bundled in
a static library for our testcase APK, our test is instrumenting against
the system server which has its own copy of SdkLevel with a higher
priority in default class loading, and that copy could be stale on
older platforms so we are seeing the NoSuchMethodError.

Instead of trying to fight with class loading, a simpler and more
reliable way is to use @SdkSuppress instead.

This also adds the explicit numeric SDK versions for the test helper
apps so that an in-development W version of the test won't fail on
stable V platform in the future.

Bug: 348576508
Flag: TEST_ONLY
Test: m gts && gts-tradefed run gts-root -m CtsRootPermissionTestCases
Test: on R/S/V platforms
Change-Id: I887e6456d4b48edb2994d5059772cbf9fcb94bc4