Allow observers to register custom executor
Add executor argument to PackageWatchdog#registerHealthObserver API.
Mitigations would be executed in the registered executors.
Add a check to ensure that observer has been registered before calling
startObservingHealth.

Note: This change is only in the module, we ignore the behaviour in
platform PackageWatchdog to keep existing functionality.

Bug: 361126781
Test: build passes
Flag: EXEMPT refactoring
Change-Id: I4f281db10010bbdfd4a88e84f3b35b07ac4506ea